Moved ShaderProperties and ShaderVersion to a models folder.
Refactored references to ShaderProperties and ShaderVersion.
